# website/static/openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vectis HTTP API",
    "version": "v1",
    "description": "Machine-readable contract for the shipped Vectis v1 HTTP API. Authentication can be disabled for local deployments; protected operations document bearer-token security because that is the production posture."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Default local vectis-api listener"
    }
  ],
...
  "paths": {
...
    "/api/v1/setup/status": {
      "get": {
        "operationId": "getSetupStatus",
        "summary": "Read setup state",
        "tags": [
          "Auth"
        ],
        "security": [],
        "x-vectis-auth-action": "setup:status",
        "responses": {
          "200": {
            "description": "Setup status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatusResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "responses": {
...
      "Error": {
        "description": "Vectis API error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIError"
            }
          }
        }
      }
    },
    "schemas": {
...
      "APIError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SetupStatusResponse": {
        "type": "object",
        "required": [
          "setup_complete",
          "auth_enabled"
        ],
        "properties": {
          "setup_complete": {
            "type": "boolean"
          },
          "auth_enabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}